+++ Runtime/Common/DataCopy.cs
@@ -7,8 +7,7 @@ namespace At.Ac.FhStp.Import3D.Common
 {
     internal static class DataCopy
     {
-        internal static Task<Nothing> CopyName(
-            INamedModel model, UnityObject obj) =>
+        internal static Task<Nothing> CopyName(INamedModel model, UnityObject obj) =>
             DoAsync(() => obj.name = model.Name);
     }
 }