+++ Runtime/Common/Instantiation.cs
@@ -6,7 +6,12 @@ namespace At.Ac.FhStp.Import3D.Common
 {
     internal static class Instantiation
     {
-        internal static Task<GameObject> MakeGameObject(string name) =>
-            CalcAsync(() => new GameObject(name));
+        internal static Task<GameObject> MakeGameObject(string name, bool active) =>
+            CalcAsync(() =>
+            {
+                var gameObject = new GameObject(name);
+                gameObject.SetActive(active);
+                return gameObject;
+            });
     }
 }