+++ Runtime/Config/GroupNodeImportConfig.cs
@@ -1,7 +1,9 @@
 ﻿namespace At.Ac.FhStp.Import3D
 {
-    public record GroupNodeImportConfig(float ScalingFactor)
+    public record GroupNodeImportConfig(
+        float ScalingFactor,
+        bool Hidden)
     {
-        public static readonly GroupNodeImportConfig Default = new GroupNodeImportConfig(1);
+        public static readonly GroupNodeImportConfig Default = new GroupNodeImportConfig(1, true);
     };
 }